Switched to a scheme where a del s.t. pn_ok == WON attempts to fix np and nnp, rather than the one s.t. del_won. Handily beating fakelflist on test_del, though deq/enq is still (unsurprisingly) worse.
The point is that this work is unnecessary if a del was helped by
help_prev(n), and pn_ok == WON iff it wasn't helped.

Note that a pn_ok == WON case won't exist if a was helped and a->p =
COMMIT won't be set in this case. lflist_enq(a, gen + 1) will have to do
some extra work to discover that it can proceed. I considered making sure
that the del_won thread sets a->p, but it's not safe to do this since it
doesn't know that np has been set and could prevent enq from making a
necessary finish_del call.