Abandoning reversed enq. It would allow spurious ABA writes on A->p.
Writing nil->p first in enq() may be possible. The problem is that it
allows A->n.st = COMMIT with A->n->p != A. And a pn write could swing p->n
= nil while nil->p != A.

Gonna see if the old scheme with a mandatory read of A->n->n is better
than a mandator write in enq.