Added .add bit. Passing -p3.
Now a->n.add and a->p.add bit are set by enq(), but not cleared when enq()
finishes. Instead, the next writes to a->n and a->p clear them. They're
used like .st = ADD in finish_enq() to avoid some reads, though bad del
patterns can make them less useful. In the worst case, where every del()
goes through deq(), finish_enq(a) must read a->n->n for every a, and must
read nil->p when deleting the next-to-last node.

Also, p.add lets you distinguish between the last phase of del and the
first phase of enq, where np != a, pn != a, p.st = RDY and n.st =
COMMIT. It'd probably be safe to treat both cases as enq() in abort_enq(),
but this avoids some writes.

Proof that finish_enq() can avoid checks if !p.add:

If a->p is written, either del(a) has finished or del(a->p) set n.st =
COMMIT (the write must be in del(a->p), help_next(a->p->p), or
help_prev(a)). So either del(a->p) or del(a) called help_next(), and thus
finish_enq() in the del(a->p) case.

Likewise, a->n.add must be cleared by del(a) or enq(a->n), both of which
resolve incomplete enq(a).