Bug: help_next() scheme fails to detect enq when a->p has been written but a->n is some old value - with .st=COMMIT. Breaks out of loop and spuriously writes a->p = COMMIT.
In fact, there's no way to distinguish between first phase enq / last
phase deq, even with .nil and .add bits. Perhaps it suffices to assume an
enq is in progress, at the cost of making a rare extra write? The response
should be to write a->n.pt = NULL, and then enq() aborts when writing
a->n.

Could it break something, to interpret del() case as enq()? Spuriously
shoot down the wrong enq()?

I think so. You see p.st == RDY but n.st == COMMIT, so you commit to write
a->n. But the del(a) finishes, and enq(a) starts anew with wrong gen. You
write a->n and spuriously shoot down that enq.

Can you write p to shoot down this phase of enq? Say you make a .pn write
to force a restart. I think not.

Gonna think a bit. Clearing .add in del would work, but I really don't
want to wind up with what I had before.