I don't see a good solution without resurrecting the ABORT state.
There's no way to tell del-final vs enq-initial states apart, and no way
to avoid spuriously jamming the "next enq" if you find the former state
and your jamming strategy is to write a->n. Well, you *could* write the
gen of the enq you were attempting to shoot down to a->n.pt, but that's
pretty awful.

Meanwhile, an ABORT state for a->p would allow for "harmless" a->p
writes. The ABORT is handled like RDY everywhere except in enq(), so at
worst you were too late to jam an enq and del(p) had to loop once on its
n->p write.