Bug: lflist_del_upd() can yield-spin forever when help_next() fails.
As mentioned, if a->p.st == ADD and a->n.st == COMMIT, it may be because
of an incomplete enq(a) or incomplete del(a), and del(a) yields until the
incomplete operation makes progress.

The problem is that I'd falsely assumed that, in the latter case, an
incomplete del(a) must have broken out of the main loop - due to winning a
p->n = n write - and will write a->p.st = COMMIT. But it could be that the
p->n write was done by a help_prev(n), and in this case every del(a) will
loop forever.

It's probably time to implement jamming and remove the yield() loop.