First draft of abort_enq() is in, but it's still disabled in favor of a yield loop. Passing -p13.
It's actually safe to yield outside the del loop. Crucially, this is
before you write n.st = COMMIT. If you find p->n != a, then some other
thread must have set n.st = COMMIT, and will exit the loop.

I've got help_next() aborting if n->p != a and n.st == ADD. The idea is
that it'll finish an enq as part of normal n->p fixup, but right now it
can spuriously write n->p if p->gen != a.gen. That is, if help_next() is
called, abort_enq() has made sure that it can safely swing n->p to finish
an enq, but *only on the current generation of a*.

It's probably time to have help_next() do a a->p.gen check upon a->n
reread. It can at least avoid this by looking at the new n.st. If n.st ==
ADD, it should fail because a must have been re-enqueued. Likewise, it can
avoid re-read upon n.st == COMMIT because, at worst, some other thread
must be deleting a after a re-enqueue. Everything it could do before
detecting the re-enqueue will be harmless.

On a related note: I've been wanting to remove the mandatory a->p re-check
in help_prev(). If pt(pn) == pt(a), its sole purpose is to avoid marking
the wrong gen as deleting. But if help_next() (and initial order of a->n
and a->p reads) already takes care of that, it's not necessary for
del(). (enq() does need to have an up-to-date picture of nil->p, but that
check could move outside help_prev()).

The best solution for these p-recheck + help_prev/next ordering questions
might be to treat the first loop iteration as special (or, indeed, just
move it out).