Bug: a good one.
Still seeing segfault when reading A in flx_of(A), with defined but unused
copy constructor. printf() legit thinks A is 0x32, though caller passes
something valid.

Disassembly looks funny too. I think flx_of() is expecting to return an
flx on the stack, via hidden argument, but its caller is expecting it in
registers and gives no hidden arg. Maybe once it's not trivially copyable,
C++ needs it on the stack?

Indeed, that's exactly what GCC's C++ ABI says:
http://mentorembedded.github.io/cxx-abi/abi.html#return-value

It's probably relevant that I silenced a bunch of silly-ish lto type
mismatch warnings over incompatible flx defs. Was fine before the advent
of copy constructors.

Dump of assembler code for function flx_of(flanchor volatile*):
0x0000000000401065 <+0>:push   %rbp
0x0000000000401066 <+1>:mov    %rsp,%rbp
0x0000000000401069 <+4>:sub    $0x10,%rsp
0x000000000040106d <+8>:mov    %rdi,-0x8(%rbp)
0x0000000000401071 <+12>:mov    %rsi,-0x10(%rbp)
0x0000000000401075 <+16>:mov    -0x10(%rbp),%rdx
0x0000000000401079 <+20>:mov    -0x8(%rbp),%rax
0x000000000040107d <+24>:mov    %rdx,%rsi
0x0000000000401080 <+27>:mov    %rax,%rdi
0x0000000000401083 <+30>:callq  0x4047f2 <flx::flx(flanchor volatile*)>
0x0000000000401088 <+35>:mov    -0x8(%rbp),%rax
0x000000000040108c <+39>:leaveq
0x000000000040108d <+40>:retq

But in the caller:
0x0000000000410558 <+647>:callq  0x406722 <linref_account_open>
0x000000000041055d <+652>:mov    -0x28(%rbp),%rax
0x0000000000410561 <+656>:add    $0x40,%rax
0x0000000000410565 <+660>:mov    %rax,%rdi
0x0000000000410568 <+663>:callq  0x401065 <flx_of(flanchor volatile*)>